setup/requirements.txt
psycopg2
pandas
mysql-connector-python
google-cloud-storage
enum34

